# app/static/swagger.yaml
swagger: "2.0"
info:
  title: Book API
  description: A social network for reading and sharing books.
  version: "1.0"
basePath: /api/v1.0
schemes:
  - http
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
security:
  - Bearer: []
paths:
...
  /books/new:
    post:
      tags:
        - Books
      summary: Create a new book
      parameters:
        - in: body
          name: book
          description: The book to create
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        201:
          description: Book created successfully
        400:
          description: Invalid input
        500:
          description: Internal server error
  /books/get/{id}:
    get:
      tags:
        - Books
      summary: Get a book by ID
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The ID of the book to retrieve
      responses:
        200:
          description: A book object
          schema:
            $ref: "#/definitions/Book"
        404:
          description: Book not found
        500:
          description: Internal server error
  /books/update/{id}:
    put:
      tags:
        - Books
      summary: Update a book by ID
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The ID of the book to update
        - in: body
          name: book
          description: The updated book object
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: Book updated successfully
        400:
          description: Invalid input
        404:
          description: Book not found
        500:
          description: Internal server error
  /books/delete/{id}:
    delete:
      tags:
        - Books
      summary: Delete a book by ID
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The ID of the book to delete
      responses:
        204:
          description: Book deleted successfully
        404:
          description: Book not found
        500:
          description: Internal server error
  /books/search:
    get:
      tags:
        - Books
      summary: Search for books based on author or book title
      parameters:
        - in: query
          name: query
          required: true
          type: string
          description: The search term to use
      responses:
        200:
          description: A list of books matching the search term
          schema:
            type: array
            items:
              $ref: "#/definitions/Book"
        400:
          description: Invalid search term
        500:
          description: Internal server error
  # user management
...
definitions:
  Book:
    type: object
    required:
      - title
      - author
      - description
      - cover_image_url
      - file_url
      - year_published
      - isbn
      - user_id
    properties:
      title:
        type: string
      author:
        type: string
      description:
        type: string
      cover_image_url:
        type: string
      file_url:
        type: string
      year_published:
        type: integer
      isbn:
        type: string
      user_id:
        type: integer
...